backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
python-dotenv==1.0.0
pymongo==4.6.1
pydantic-settings==2.1.0
email-validator>=2.1.1
certifi>=2024.2.2
